+++ Resources/JS/test.js
@@ -1,0 +1,1 @@
+$('#back .specs-table span').text("Test text");
